Created adapter to calories recyclerView and interface to calories fragment and calories item

+++ app/src/main/java/com/example/caloriesapi/domain/entity/CaloriesItem.kt
@@ -2,8 +2,8 @@ package com.example.caloriesapi.domain.entity
 
 data class CaloriesItem(
     val label: String,
-    val energy: Int,
-    val protein: Int,
-    val fat: Int,
-    val carbohydrate: Int
+    val energy: Double,
+    val protein: Double,
+    val fat: Double,
+    val carbohydrate: Double
 )
